test-requirements.txt
flake8
paramiko
libvirt-python
mock
nose
